+++ .github/scripts/utility.mts
@@ -3,20 +3,11 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
+// NOTE: このファイルはworkflow上でバックエンドからも参照されるため、side effectがあってはならない
+
 import { spawn } from 'node:child_process';
 import { promises as fs } from 'node:fs';
 import path from 'node:path';
-
-export const heapSnapshotCategories = [
-	'Total',
-	'Code',
-	'Strings',
-	'JS arrays',
-	'Typed arrays',
-	'System objects',
-	'Other JS objects',
-	'Other non-JS objects',
-] as const;
 
 export function median(values: number[]) {
 	const sorted = values.toSorted((a, b) => a - b);
@@ -30,6 +21,40 @@ export function mad(values: number[]) {
 
 	const center = median(values);
 	return median(values.map(value => Math.abs(value - center)));
+}
+
+function getSamplesByRound<T extends { round: number; }[]>(samples: T) {
+	const samplesByRound = new Map<number, T[number]>();
+	for (const sample of samples) {
+		if (sample.round <= 0) continue;
+		samplesByRound.set(sample.round, sample);
+	}
+	return samplesByRound;
+}
+
+export function pairedDeltaSummary<T extends { round: number; }[]>(baseSamples: T, headSamples: T, getValue: (sample: T[number]) => number | null) {
+	const baseSamplesByRound = getSamplesByRound(baseSamples);
+	const headSamplesByRound = getSamplesByRound(headSamples);
+	const values = [];
+
+	for (const [round, baseSample] of baseSamplesByRound) {
+		const headSample = headSamplesByRound.get(round);
+		if (headSample == null) continue;
+
+		const baseValue = getValue(baseSample);
+		const headValue = getValue(headSample);
+		if (baseValue == null || headValue == null) continue;
+
+		values.push(headValue - baseValue);
+	}
+
+	return {
+		median: median(values),
+		mad: mad(values),
+		min: Math.min(...values),
+		max: Math.max(...values),
+		samples: values.length,
+	};
 }
 
 export function normalizePath(filePath: string) {
